requirements.txt
aiogram==2.19
aiohttp==3.8.1
aiosignal==1.2.0
aresponses==2.1.5
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.4.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
frozenlist==1.3.0
idna==3.3
iniconfig==1.1.1
multidict==6.0.2
packaging==21.3
Pillow==9.0.0
pluggy==1.0.0
pony==0.7.16
psycopg2==2.9.3
py==1.11.0
pyparsing==3.0.7
pytest==6.2.5
pytest-asyncio==0.18.3
pytz==2021.3
requests==2.27.1
toml==0.10.2
urllib3==1.26.9
yarl==1.7.2
